필수값 막기
